refactor: split large components, deduplicate shared logic
Go backend:
- repository/events.go: extract topUTMColumn helper — TopUTMSources,
  TopUTMCampaigns, TopUTMMediums now delegate instead of repeating
  the same query/scan pattern
- api/middleware.go: extract firstXFF helper to remove duplicated
  X-Forwarded-For first-hop parsing

Frontend — component splits (page line counts after):
- Settings.tsx 1121→354: ProjectSettings + ApiKeySettings sub-components
- Flags.tsx 1217→677: FlagFormModal + TargetingRuleEditor + shared helpers
- Funnels.tsx 2537→734: ImplementationSnippet, StepFilterEditor,
  CreateFunnelModal, EditFunnelModal, ScopeToggle, SegmentManager,
  constants extracted to components/funnels/
- Insights.tsx 645→203: WidgetCard, AddWidgetModal, VisitorBreakdown
- Shell.tsx 549→130: TopNav, BottomTabBar, MoreSheet, ShellStyles

Shared infrastructure:
- web/src/lib/theme.ts: single C color palette, replaces identical
  const C blocks in Dashboard, Insights, Live, Landing, Login, Shell
- Dashboard.tsx: calcTrend() helper replaces duplicate IIFEs; local
  Skeleton removed in favour of components/ui/Skeleton
