+++ web/src/lib/theme.ts
@@ -1,0 +1,10 @@
+export const C = {
+  bg: '#0f1117',
+  surface: '#1a1d27',
+  border: '#2a2d3a',
+  amber: '#f59e0b',
+  text: '#e2e8f0',
+  muted: '#94a3b8',
+  success: '#10b981',
+  error: '#ef4444',
+} as const
