+++ web/src/pages/Live.tsx
@@ -6,16 +6,7 @@ import Shell from '../components/shell/Shell'
 import { api } from '../lib/api'
 import { useProjects } from '../lib/projects'
 import { Skeleton } from '../components/ui/Skeleton'
-
-const C = {
-  bg: '#0f1117',
-  surface: '#1a1d27',
-  border: '#2a2d3a',
-  amber: '#f59e0b',
-  text: '#e2e8f0',
-  muted: '#94a3b8',
-  success: '#10b981',
-}
+import { C } from '../lib/theme'
 
 interface LiveEvent {
   id: string
